Cleanup: remove fs_stream_pipe.{h,cpp}
