+++ be/src/exec/vectorized/json_scanner.cpp
@@ -694,8 +694,8 @@ Status JsonReader::_read_and_parse_json() {
     {
         SCOPED_RAW_TIMER(&_counter->file_read_ns);
         // For efficiency reasons, simdjson requires a string with a few bytes (simdjson::SIMDJSON_PADDING) at the end.
-        RETURN_IF_ERROR(stream_file->read_one_message(&_parser_buf, &_parser_buf_cap, &_parser_buf_sz,
-                                                      simdjson::SIMDJSON_PADDING));
+        RETURN_IF_ERROR(stream_file->pipe()->read_one_message(&_parser_buf, &_parser_buf_cap, &_parser_buf_sz,
+                                                              simdjson::SIMDJSON_PADDING));
         if (_parser_buf_sz == 0) {
             return Status::EndOfFile("EOF of reading file");
         }
