+++ be/src/fs/CMakeLists.txt
@@ -29,7 +29,6 @@ set(EXEC_FILES
     fs.cpp
     fs_posix.cpp
     fs_util.cpp
-    fs_stream_pipe.cpp
     fs_broker.cpp
     fs_memory.cpp
     fs_s3.cpp
