+++ be/src/runtime/stream_load/stream_load_pipe.h
@@ -25,6 +25,7 @@
 #include <deque>
 #include <mutex>
 
+#include "io/input_stream.h"
 #include "runtime/message_body_sink.h"
 #include "util/bit_util.h"
 #include "util/byte_buffer.h"
@@ -212,6 +213,22 @@ private:
     ByteBufferPtr _write_buf;
     ByteBufferPtr _read_buf;
     Status _err_st = Status::OK();
+};
+
+// TODO: Make `StreamLoadPipe` as a derived class of `io::InputStream`.
+class StreamLoadPipeInputStream : public io::InputStream {
+public:
+    explicit StreamLoadPipeInputStream(std::shared_ptr<StreamLoadPipe> file);
+    ~StreamLoadPipeInputStream() override;
+
+    StatusOr<int64_t> read(void* data, int64_t size) override;
+
+    Status skip(int64_t n) override;
+
+    std::shared_ptr<StreamLoadPipe> pipe() { return _pipe; }
+
+private:
+    std::shared_ptr<StreamLoadPipe> _pipe;
 };
 
 } // namespace starrocks
